chore: renamed plugin-input.ts
