+++ src/main.ts
@@ -2,7 +2,7 @@ import { createActionsPlugin } from "@ubiquity-os/plugin-sdk";
 import { createAdapters } from "./adapters";
 import { SupportedEvents } from "./types/context";
 import { Env, envSchema } from "./types/env";
-import { PluginSettings, pluginSettingsSchema } from "./types/plugin-inputs";
+import { PluginSettings, pluginSettingsSchema } from "./types/plugin-input";
 import { Command } from "./types/command";
 import { plugin } from "./plugin";
 import { LogLevel } from "@ubiquity-os/ubiquity-os-logger";
