+++ src/types/index.ts
@@ -1,3 +1,3 @@
 export * from "./context";
 export * from "./env";
-export * from "./plugin-inputs";
+export * from "./plugin-input";
